+++ plugins/org.tetrabox.minijava.xtext/src-gen/org/tetrabox/minijava/xtext/miniJava/impl/IntConstantImpl.java
@@ -35,7 +35,7 @@ public class IntConstantImpl extends ExpressionImpl implements IntConstant
    * @generated
    * @ordered
    */
-  protected static final int VALUE_EDEFAULT = 0;
+  protected static final double VALUE_EDEFAULT = 0.0;
 
   /**
    * The cached value of the '{@link #getValue() <em>Value</em>}' attribute.
@@ -45,7 +45,7 @@ public class IntConstantImpl extends ExpressionImpl implements IntConstant
    * @generated
    * @ordered
    */
-  protected int value = VALUE_EDEFAULT;
+  protected double value = VALUE_EDEFAULT;
 
   /**
    * <!-- begin-user-doc -->
@@ -73,7 +73,7 @@ public class IntConstantImpl extends ExpressionImpl implements IntConstant
    * <!-- end-user-doc -->
    * @generated
    */
-  public int getValue()
+  public double getValue()
   {
     return value;
   }
@@ -83,9 +83,9 @@ public class IntConstantImpl extends ExpressionImpl implements IntConstant
    * <!-- end-user-doc -->
    * @generated
    */
-  public void setValue(int newValue)
+  public void setValue(double newValue)
   {
-    int oldValue = value;
+    double oldValue = value;
     value = newValue;
     if (eNotificationRequired())
       eNotify(new ENotificationImpl(this, Notification.SET, MiniJavaPackage.INT_CONSTANT__VALUE, oldValue, value));
@@ -118,7 +118,7 @@ public class IntConstantImpl extends ExpressionImpl implements IntConstant
     switch (featureID)
     {
       case MiniJavaPackage.INT_CONSTANT__VALUE:
-        setValue((Integer)newValue);
+        setValue((Double)newValue);
         return;
     }
     super.eSet(featureID, newValue);
